requirements.txt
flask==2.0.1
pymysql==1.0.2
cryptography==3.2
PyYAML==5.4.1
requests==2.26.0
pyjwt==1.3.0